backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
websockets==12.0
groq==0.8.0
edge-tts==6.1.10
python-multipart==0.0.9
python-dotenv==1.0.1
